gauges are now double as allowed by statsd, counters and timers are prefered to be ints

+++ src/StatsdClient/Metrics.cs
@@ -30,7 +30,7 @@ namespace StatsdClient
             _statsD.Send<Statsd.Counting>(BuildNamespacedStatName(statName), value, sampleRate);
 		}
 
-		public static void Gauge(string statName, int value)
+		public static void Gauge(string statName, double value)
 		{
 			if (_statsD == null)
 			{
